# openapi.yaml
openapi: 3.0.3
info:
  title: Rani Beauty Clinic API
  version: 1.0.0
  description: |
    Baseline API contract for production route surfaces.
    This spec is CI-validated and intended to evolve with the codebase.
  license:
    name: Proprietary
    url: https://rani-beauty-clinic.vercel.app
servers:
  - url: https://rani-beauty-clinic.vercel.app
    description: Production
...
security:
  - SessionCookie: []
paths:
...
  /api/patient/profile:
    get: { operationId: getPatientProfile, summary: Patient profile, tags: [Patient], security: [{ SessionCookie: [] }], responses: { "200": { description: Profile data }, "401": { $ref: "#/components/responses/Unauthorized" } } }
    patch: { operationId: updatePatientProfile, summary: Update patient profile, tags: [Patient], security: [{ SessionCookie: [] }], requestBody: { required: true, content: { application/json: { schema: { type: object } } } }, responses: { "200": { description: Profile updated }, "400": { $ref: "#/components/responses/BadRequest" }, "401": { $ref: "#/components/responses/Unauthorized" } } }
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: rani-session
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...